+++ packages-exp/installations-compat/src/testing/util.ts
@@ -37,8 +37,7 @@ export function getFakeApp(): FirebaseApp {
     },
     automaticDataCollectionEnabled: true,
     delete: async () => {},
-    installations: (() => null as unknown) as any,
-    _delegate: {} as any
+    installations: (() => null as unknown) as any
   };
 }