+++ packages/installations/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @firebase/installations
 
+## 0.4.23
+
+### Patch Changes
+
+- Updated dependencies [[`de5f90501`](https://github.com/firebase/firebase-js-sdk/commit/de5f9050137acc9ed1490082e5aa429b5de3cb2a)]:
+  - @firebase/util@0.4.1
+  - @firebase/component@0.3.1
+
 ## 0.4.22
 
 ### Patch Changes